requirements.txt
torch
torchvision
numpy
matplotlib
pyrender
trimesh
pillow